Adds new phrase config option homerow-only; add log read/write
For example, may be interesting to compare full qwerty availability (phrases_default.txt, ~18.5 bits/char) with homerow-only (phrases_homerow-only.txt, ~10 bits/char) since we should not have to move as far we should be able to strike the characters faster in `homerow-only.txt` version of phrases.

Binary file logging also added. In general the idea is to have a single `tic()` in a main `.m` script. This is passed to classes like `typewriter.Prompter` or `cursor.Cursor` so that they can then log accurate binary timestamps to respective `.typewriterdata` or `.cursordata` binaries. Because the alignment tic is from a single source, post-hoc alignment should therefore be straightforward.